Add files via upload
The note file in the project repository is a file that contains updates and notes about the current status and progress of the project. Its purpose is to provide a place where developers and contributors can share important information about the project, such as changes to the project's goals or updates to the codebase. It serves as a useful reference point for anyone who wants to get up to speed on the project quickly, and it helps to ensure that everyone involved in the project is on the same page.